+++ src/neuroconv/datainterfaces/ecephys/openephys/openephysbinarydatainterface.py
